+++ bin/bossStateAPI.js
@@ -4,7 +4,7 @@ require('dotenv').config()
 
 const https = require('https')
 const ExternalSystemException = require('../src/boss/exception/ExternalSystemException')
-const initRaidRecord = require('../middlewares/cache')
+const initRaidRecord = require('../src/middlewares/cache')
 
 const options = {
   hostname: process.env.BOSS_STATE_URL,
